fix(replay): Fix canvas recordings crashing the browser
Fixes an issue with canvas replayer where it would crash the browser for replays with a large amount of canvas frames. This happened because we were attempting to preload every canvas event on load.

This PR changes the canvas player to only preload up to 50 events in the future and will clear preloaded events not within 20 seconds of the current player time.

Closes https://github.com/getsentry/sentry/issues/65792